+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.5.6] 2024-07-05
+
 - Add LPSPI low-level clock configuration APIs.
 - Add LPSPI `set_peripheral_enable` to configure the driver as a SPI peripheral.
 
@@ -316,7 +318,8 @@ The release includes 0.3.1 fixes.
 
 Prior releases were not tracked with a changelog entry.
 
-[Unreleased]: https://github.com/imxrt-rs/imxrt-hal/compare/0.5.5...main
+[Unreleased]: https://github.com/imxrt-rs/imxrt-hal/compare/0.5.6...main
+[0.5.6]: https://github.com/imxrt-rs/imxrt-rs/compare/0.5.5...0.5.6
 [0.5.5]: https://github.com/imxrt-rs/imxrt-rs/compare/0.5.4...0.5.5
 [0.5.4]: https://github.com/imxrt-rs/imxrt-rs/compare/0.5.3...0.5.4
 [0.5.3]: https://github.com/imxrt-rs/imxrt-rs/compare/0.5.2...0.5.3